integration: 1) changed almost every file. The integration code can now hold monomials with rational (RationalNLT vs ZZ) coefficients. 2) Added a test integration maple file that should be treated like a lib. The lib's "make polynomial" function differs from the ones defined in other files in a few  ways: it has rational coefficients, negative coefficients, monomials of degree from 1 to max_degree, and a random number of monomials.
valuation:
1) Moved the old functions in the driver to valuation.cpp and placed them in a namespace. Driver.cpp looks more like a driver now. 
2) testIntegrationHyperrectangles.cpp and integrateHyperrectangle.mpl are a driver for computing the integrals over hyper-rectangles and checking the computed answer with the real one. You can run the test by typing: maple -q valuation/integrateHyperrectangle.mpl. I should add a shell script that calls this line for the user.
3) Driver.cpp now computes the integral over rational-vertex polytopes by dilation. The current default is integration. Type ValuationComputation --help to see the help menu. Polynomials can be given by the user from the termal in [[coef, [exps]]+ ] syntax or as the first line of a file, and the file name is passed on the command line. Currently I do no extra parsing of the polynomial for syntax errors. 
4) The PolytopeValuation class still has the old integration code commented out (when the integration code could only use integer coefficients), this should be removed next.

+++ code/latte/testEhrhart/BuildRandomPolytope.cpp
@@ -21,6 +21,8 @@ BuildRandomPolytope::BuildRandomPolytope(int ambient_dim)
 	integerPoints = true;
 	probNegative = .5;
 	numAffineHull = 0;
+	savePolymakeFile = false; //don't save the file. Delete it in the deconstructor.
+	saveLatteFile = false;
 
 }//buildRandomPolytope
 
@@ -28,14 +30,20 @@ BuildRandomPolytope::BuildRandomPolytope(int ambient_dim)
 BuildRandomPolytope::~BuildRandomPolytope()
 {
 	stringstream command;
-	command << "rm -f " << fileName.c_str();
-	cout << "Deleting file " << command.str().c_str() << endl;
-	system(command.str().c_str());
-	command.str("");
-	command.clear();
-	command << "rm -f " << latteFile.c_str();
-	cout << "Deleting file " << command.str().c_str() << endl;
-	system(command.str().c_str());
+	if ( savePolymakeFile == false)
+	{
+		command << "rm -f " << fileName.c_str();
+		cout << "Deleting file " << command.str().c_str() << endl;
+		system(command.str().c_str());
+		command.str("");
+	}
+	if (saveLatteFile == false)
+	{
+		command.clear();
+		command << "rm -f " << latteFile.c_str();
+		cout << "Deleting file " << command.str().c_str() << endl;
+		system(command.str().c_str());
+	}
 }//deconstructor.
 
 /**
@@ -300,6 +308,24 @@ void BuildRandomPolytope::printFacetEquationsForLattE()
 	}//if there are affine hulls!
 	file.close();
 }//printFacetEuationsForLatte()
+
+/**
+ * Changes the save state of the latte file.
+ */
+void BuildRandomPolytope::saveTheLatteFile(bool save)
+{
+	saveLatteFile = save;
+}
+
+/**
+ * Changes the save state of the polymake file.
+ */
+
+void BuildRandomPolytope::saveThePolymakeFile(bool save)
+{
+	savePolymakeFile = save;
+}
+
 
 /**
  * This string is passed to ehrhart2 and is printed in the log files to make things easier to read.
